algunas actualizaciones malas
